chore: rename file from test to recipeDTOs
also remove unused variable
